Linked view_selected_comments with GET source

+++ admin/includes/view_selected_comments.php
@@ -17,55 +17,58 @@
 	<tbody>
 							
 		<?php
-		// Collect each row from database
-		$query = "SELECT * FROM comments ";
-		$select_comments = mysqli_query($connection, $query);
-		while($row = mysqli_fetch_assoc($select_comments)){
-			$comment_id = $row['comment_id'];
-			$comment_post_id = $row['comment_post_id'];
-			$comment_author = $row['comment_author'];
-			$comment_email = $row['comment_email'];
-			$comment_content = $row['comment_content'];
-			$comment_status = $row['comment_status'];
-			$comment_date = $row['comment_date'];
+		if(isset($_GET['p_id'])){
+			$post_id = $_GET['p_id'];
+		
+			// Collect each row from database
+			$query = "SELECT * FROM comments WHERE comment_post_id = $post_id ";
+			$select_comments = mysqli_query($connection, $query);
+			while($row = mysqli_fetch_assoc($select_comments)){
+				$comment_id = $row['comment_id'];
+				$comment_post_id = $row['comment_post_id'];
+				$comment_author = $row['comment_author'];
+				$comment_email = $row['comment_email'];
+				$comment_content = $row['comment_content'];
+				$comment_status = $row['comment_status'];
+				$comment_date = $row['comment_date'];
 
-			echo "<tr>";
-			echo "<td> $comment_id </td>";
-			echo "<td> $comment_author </td>";
-			echo "<td> $comment_content </td>";
+				echo "<tr>";
+				echo "<td> $comment_id </td>";
+				echo "<td> $comment_author </td>";
+				echo "<td> $comment_content </td>";
 
-	/*		// Retreive id and category for database
-			$query = "SELECT * FROM categories WHERE cat_id = {$post_Category} ";
-			$cat_info_query =  mysqli_query($connection, $query);
+		/*		// Retreive id and category for database
+				$query = "SELECT * FROM categories WHERE cat_id = {$post_Category} ";
+				$cat_info_query =  mysqli_query($connection, $query);
 
-			while($row = mysqli_fetch_assoc($cat_info_query)){
-				$cat_name = $row['cat_title'];
-					echo "<td> {$cat_name} </td>";
-				
+				while($row = mysqli_fetch_assoc($cat_info_query)){
+					$cat_name = $row['cat_title'];
+						echo "<td> {$cat_name} </td>";
+
+				}
+		*/
+
+				echo "<td> $comment_email </td>";
+				echo "<td> $comment_status </td>";
+
+				// Relational Database Tables: comments , posts
+				$query = "SELECT * FROM posts WHERE post_id = $comment_post_id ";
+				$select_post_id_query = mysqli_query($connection, $query);
+				while($row = mysqli_fetch_assoc($select_post_id_query)){
+
+					$post_id = $row['post_id'];
+					$post_title = $row['post_title'];
+
+					echo "<td><a href='../post.php?p_id=$post_id'>$post_title</a> </td>";
+				}
+
+				echo "<td> $comment_date </td>";
+				echo "<td><a href='comments.php?approve=$comment_id'>Approve</a></td>";
+				echo "<td><a href='comments.php?unapprove=$comment_id'>Unapprove</a></td>";
+				echo "<td><a href='comments.php?delete=$comment_id'>Delete</a></td>";
+				echo "</tr>";
 			}
-	*/
-
-			echo "<td> $comment_email </td>";
-			echo "<td> $comment_status </td>";
-			
-			// Relational Database Tables: comments , posts
-			$query = "SELECT * FROM posts WHERE post_id = $comment_post_id ";
-			$select_post_id_query = mysqli_query($connection, $query);
-			while($row = mysqli_fetch_assoc($select_post_id_query)){
-				
-				$post_id = $row['post_id'];
-				$post_title = $row['post_title'];
-				
-				echo "<td><a href='../post.php?p_id=$post_id'>$post_title</a> </td>";
-			}
-			
-			echo "<td> $comment_date </td>";
-			echo "<td><a href='comments.php?approve=$comment_id'>Approve</a></td>";
-			echo "<td><a href='comments.php?unapprove=$comment_id'>Unapprove</a></td>";
-			echo "<td><a href='comments.php?delete=$comment_id'>Delete</a></td>";
-			echo "</tr>";
 		}
-
 		?>
 							
 	</tbody>
